niffler-e-2-e-tests-python/requirements.txt
pytest
requests
python-dotenv

psycopg2-binary
sqlalchemy
sqlmodel
playwright==1.46.0
faker